Underpriced transactions go down the gas-bump path
Since underpriced transactions indicate the submitted transaction for
that nonce has too low a gas price, entering the gas-bump loop allows
the gas bumping to get the transaction to an appropriate price.

To do this, _track_tx_result can now operate without a tx_id. This is
because, in the case of underpriced transactions, no tx_id is available
to pass along. A lack of tx_id is treated as a transaction needing
retry.