Add logging and retries when boosting gas to match existing tx
When the maintainer restarts, its transaction resubmissions start at 1
tick and have to catch up to the number of ticks of the existing pending
transactions on chain, if any. In these cases, the new gas price can
end up below the latest gas price. For these situations, the gas price
is now boosted until it either reaches MAX_GAS_PRICE or exceeds the
latest gas price.

Additionally, irrespective of how the gas bump path is handled, the path
through it is logged.