Bump gas per wait unit more aggressively
Right now the wait unit is the product of how many pending transactions
are ahead of this one (generally only revelevant during relay catch-up)
and the number of ticks this transaction has gone unmined (with one tick
= 30 seconds). We were cranking up by 20% of this wait unit multipled by
default gas every tick, but this commit adjusts it to crank up by 50% of
the wait unit * default gas. For a linear wait unit increase, which is
the standard expected behavior, this yields:

  Time  | Gas Price
--------------------
   0    | 100 gwei
  30s   | 150 gwei
  60s   | 200 gwei
 1m30s  | 250 gwei
  2m    | 350 gwei
 2m30s  | 400 gwei
  3m    | 450 gwei
 3m30s  | 500 gwei
  4m    | 550 gwei

With recent gas prices, this should mean relay information should land
within 4-5 minutes when gas prices are high, but much faster and cheaper
when prices are lower. If this feels too long or gas prices become more
volatile still, the numbers can be adjusted.