Configurable Gas Price details
We let DEFAULT_GAS_PRICE used for transaction submission and
MAX_GAS_PRICE used as a limit for gas price bumps to be configurable
instead of hardcoding them.

The default values will be set the same as previously hardcoded:
DEFAULT_GAS_PRICE = 100 Gwei
MAX_GAS_PRICE = 600 Gwei

Having it configurable is better from multi-environment perspective, as
it doesn't make sense to start with 100 Gwei on Ropsten.

It also gives possibility to easily tweak settings for mainnet without
the need to rebuild the code.

Following properties can be used to configure it:
DEFAULT_GAS_PRICE_GWEI and MAX_GAS_PRICE_GWEI. Note that configured
values are expected to be provided in Gwei (not wei!).